fix: resolve test failures and move device-specific tests
- Move NPU-dependent tests to tests/npu/ (dispatch, pin_memory)
- Move CUDA-dependent tests to tests/cuda/ (device transfer)
- Fix make_tensor scalar shape handling for numpy scalar
- Add contract conftest.py to restore registry state between tests
  (fixes 26 contract test failures from registry pollution)
