+++ tests/cpu/test_dispatch_resolution.py
@@ -12,13 +12,6 @@ def test_dispatch_prefers_meta_when_input_meta():
     assert c.device.type == "meta"
 
 
-def test_dispatch_prefers_npu_over_cpu():
-    a = torch.ones((2,), device="npu")
-    b = torch.ones((2,), device="npu")
-    c = torch.add(a, b)
-    assert c.device.type == "npu"
-
-
 def test_dispatch_rejects_cross_npu_device_index():
     class _FakeTensor:
         def __init__(self, dev):
